Update filtering.md
I suggest some additional explanation of what happens when __EnableDeferredFiltering__ is true. Is the search as you type activity stopped or are keystrokes still stored until lost focus/enter or tab, and then after a tab does the filtering activity still happen? Also, is there an event called lost focus/enter? If not, I suggest you include the real names of the events.