Update disable-ieditableobject-support.md
Several small edits and several comments about better SEO inline. Please address and then remove my comments.

I suggested to Marin on the AJAX team that it might help all of your teams to have a private repository of capitalization and style rules (like whether a value should be written as **True** or false) in GitHub that we all have access to. It would be GREAT if your teams built a sort of style check tool just to check capitalization and style rules like this before articles get published in GitHub. It could immensely help all of the Telerik teams that write documentation if everyone used it. I use a tool in Word that flags many style rule issues (but I don't run it on Telerik documentation because there isn't enough consistency to enforce and it doesn't like Markdown syntax). 
Here is an idea: if Telerik created a great style and capitalization tool that works for GitHub that you build for your own documentation, you could sell the tool to customers. Tools like that don't get out of date fast (grammar rules are very slow to change, but if you made a dictionary that allowed customers to add their own terms and style rules, that would be the best option) and you wouldn't have to do much marketing -- it's an easy add-on purchase for your current customer lists.  You could even give basic version away for a while until it has good name recognition and then sell a Pro version.