Eriks-edits
There are a few comments in here plus copy changes.

I didn't mark it for change, but it sounds odd to me to use the preposition "on" to describe the content shown in a figure. I am much more familiar with "in the figure" or "in Figure 4" than "on the figure" or "on Figure 4". The same would be true for "on a control" or "on the ScheduleView" where I think it is correct to use "in" rather than "on". Your article isn't the only one I see this in, so please bring it up at a team meeting to discuss. The writers are not consistent from article to article. This affects the the XAML and AJAX teams so whatever you decide to use (I prefer in) then I'll make those changes to all articles going forward.