Update common-localization.md
Some copy editing. Please review the comments.
The first part of the article I struggled to figure out the difference between the built-in localization mechanism and the custom localization manager. They sound very similar and that is confusing. Please explain, with an example (not necessarily a code example) when/why you would use one versus the other. I read the article a few times and I do not know understand when one is better than the other.