copyediting
Copy editing changes only. No comments to address. However, why is ActivePaneChangedEventArgs (at the top of the article) indented one level more than the other events in the list. It stands out as odd to me. I suggest it gets moved left so that the events in that list are all aligned the same.