Many comments and edits
This article had a lot of edits and comments to review. The introduction was confusing and the numbering of examples didn't seem to match what I'm seeing in other help articles. I'd like to review this again after you've made fixes to it.

It also was not consistent about using the Telerik style for lists. Please only use numbered lists for step-by-step instructions.