Update backgrounds-borders.md
Line 13: Do we mean to say, "The controls in the RadGauge suite"?

What do we mean by "expose" do we mean "enable" or "have"? (See also lines 22 and 38.)

"can be used to customize their backgrounds and borders" What does "their" refer to--the controls' backgrounds and borders?

I have deleted periods from sentence fragments (wherever the sentences are not complete sentences). If any one sentence in a bulleted list is complete, all the sentences can have periods at the end.