Update xaml-vs-noxaml.md
I caught what looks like a minor typo in the NoXaml assms paragraph - the second sentence compares NoXaml to itself; should it be Xaml?

The NoXaml dlls have the same code as the Xaml ones, but they don't contain any styles and templates. This makes them __smaller in size__ compared to __Xaml__, but they require a bit more effort to set up.